Initial try at letting users select preferred locale
This is a small step in internationalizing amo-cli. These changes
specifically allow the user to select their preferred locale, which is used
in for the `lang` query parameter when making GET requests aginst the AMO API.
This is beneficial because the returned request will have add-on `name`, `description`,
and other fields translated in the selected locale if supplied by add-on authors.